requirements.txt
backcall==0.1.0
backports.csv==1.0.7
colorlog==4.0.2
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-autocomplete-light==3.3.4
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-glrm==1.1.3
django-import-export==1.2.0
django-querysetsequence==0.11
django-tables2==2.0.6
django-xhtml2pdf==0.0.4
et-xmlfile==1.0.1
html5lib==1.0.1
ipython==7.4.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.13.3
odfpy==1.4.0
openpyxl==2.6.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
PyPDF2==1.26.0
pytz==2019.1
PyYAML==5.1
reportlab==3.5.20
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
